resolve ipv4 addr from hostname; refactor web_assets
