+++ crates/api-ui/src/web_assets/mod.rs
@@ -1,6 +1,9 @@
 pub mod config;
 pub mod error;
 pub mod handler;
-pub mod server;
+pub mod router;
+#[cfg(test)]
+pub mod tests;
 
 pub use error::Error;
+pub use router::web_assets_app;
