+++ crates/embucket-seed/README.md
@@ -8,7 +8,7 @@ Implements seeding data for embucket database. Supports 3 levels of seeding: Min
 embucket-seed --server-address '127.0.0.1:3000' --auth-user embucket --auth-password embucket --seed-variant typical
 
 # or build & run with cargo run
-cargo embucket-seed -- --server-address '127.0.0.1:3000' --auth-user embucket --auth-password embucket --seed-variant typical
+cargo embucket-seed
 ```
 
 ## Updating seed templates
